requirements.txt
python-dotenv==1.0.1
flask==3.0.2
alerts-in-ua.py==1.2.6